// src/presentation/Enums/CristalesEnum.ts
export enum CristalesEnum {
  codigo = 1,
  ubicacion = 2,
  stock_minimo = 3,
  stock_reservado = 4,
  stock_disponible = 5,
  marca_id = 6,
  marca = 7,
  proveedor_id = 8,
  proveedor = 9,
  diseno_id = 10,
  diseno = 11,
  indice_id = 12,
  indice = 13,
  material_id = 14,
  material = 15,
  color_id = 16,
  color = 17,
  tratamiento_id = 18,
  tratamiento = 19,
  diametro = 20,
  esferico = 21,
  cilindrico = 22,
  codigo_fab_1 = 23,
  codigo_fab_2 = 24,
  codigo_fab_3 = 25,
  codigo_fab_4 = 26,
}